Get interface through packet_source() bif instead of cluster layout
Zeek 3.1 introduced the packet_source() bif. It can be leveraged to reliably
get the live interface of a worker. Whether cluster-layout.zeek / Cluster::nodes
provides the interface field is not guaranteed.

This change allows to use this package in environments where the
Cluster::Node$interface field is not populated. Further, it adds
the interface for a simple "zeek -i eth0" invocation which seems nice.

Technically, the only reason this depends on Cluster now is the fallback
to the "unknown" interface and I'm not quite sure that's actually useful
for when logs are generated on proxies or managers.